AP_Math: split out double precision location functions
this allows ALLOW_DOUBLE_MATH_FUNCTIONS to be used
