+++ libraries/AP_Math/examples/location/location.cpp
@@ -2,6 +2,8 @@
 // Unit tests for the AP_Math polygon code
 //
 
+#define ALLOW_DOUBLE_MATH_FUNCTIONS
+
 #include <AP_HAL/AP_HAL.h>
 #include <AP_Math/AP_Math.h>
